Duplicate code detected at lines 262-347 (86 lines total)
First 3 lines:
262:             msg = f"Failed to create commit: {e}"
263:             raise VCSOperationError(msg) from e
264:

... (80 lines omitted) ...

Last 3 lines:
345:                 + "\n\nNote: Untracked files are OK and won't block processing."
346:             )
347:             raise DirtyWorkingDirectoryError(msg)

This code is duplicated in 1 other location(s):
  - vh_alexei_eleusis_gmail_com_mercurial_251217-1548:src/vibe_heal/vcs/mercurial/manager.py (lines 310-395)
